fix(bp-analysis): address code review findings (Critical #1, #2 + Important #5, #6 + Minor #11, #14)
Critical #1: Add SelectedNodes case in serialization analysis branch — was silently
producing empty graphs JSON (fell through to default:break in scope switch).

Critical #2: Replace raw `this` lambda capture in OnAnalyzeClicked with TWeakPtr<>
(WeakSelf) to prevent use-after-free if tab is destroyed before async callback.

Important #5: Replace O(n²) GUID reverse-lookup in TakeOwnershipOfClonedGraphs with
pre-built TMap<FGuid, int32> for O(n) NodeGraphNames population.

Important #6: Reset BlueprintCompiledHandle after delegate removal in destructor to
prevent stale handle from staying valid when Blueprint is already unloaded.

Minor #11: Fix integer division in DurationMs display — cast to (double) before
dividing by 1000.0 to preserve sub-second precision.

Minor #14: Add EMessageSeverity enum value comment (CriticalError=0, Error=1, Warning=3)
to clarify the <= comparison ordering.

+++ Source/CortexFrontend/Private/Widgets/SCortexAnalysisChat.cpp
@@ -237,7 +237,7 @@ void SCortexAnalysisChat::CollapseStatusMessages(const FCortexTurnResult& Result
 	else
 	{
 		SummaryEntry->Text = FString::Printf(TEXT("Analysis completed in %.1fs"),
-			Result.DurationMs / 1000.0);
+			(double)Result.DurationMs / 1000.0);
 	}
 
 	TSharedPtr<FCortexChatDisplayRow> SummaryRow = MakeShared<FCortexChatDisplayRow>();
